Coding part Finished for project2draft.qmd (#1)
* Update project2draft.qmd

Rewrite the conclusion for TESTING only.

* Update project2draft.qmd

The model fits all the variables, but the result is not really great.

* Update project2draft.qmd

Added the part of deleting outliers and changed the model again, but the formula and some analysis results have not been changed.

* Update project2draft.qmd

Regard the Model 3 as the final model. Interpretation and conclusion are left to you guys~ ☺️